==> Notebooks/1-loading_raw_csv.ipynb <==
import pandas as pd 

df = pd.read_csv("D:/ml-aws/data/raw/HouseTS.csv")
df.head()


df['date'] = pd.to_datetime(df['date'])
print(df['date'].dt.year.value_counts())

df = df.sort_values("date")

# Define cutoff dates
cutoff_date_eval = "2020-01-01"   # validation starts
cutoff_date_holdout = "2022-01-01"  # holdout starts
# Train: before 2020
train_df = df[df["date"] < cutoff_date_eval]

# Validation/Eval: 2020–2021
eval_df = df[(df["date"] >= cutoff_date_eval) & (df["date"] < cutoff_date_holdout)]

# Holdout: 2022–2023
holdout_df = df[df["date"] >= cutoff_date_holdout]

print("Train shape:", train_df.shape)
print("Eval shape:", eval_df.shape)
print("Holdout shape:", holdout_df.shape)

train_df.to_csv("D:/ml-aws/data/raw/train.csv", index=False)
eval_df.to_csv("D:/ml-aws/data/raw/eval.csv", index=False)
holdout_df.to_csv("D:/ml-aws/data/raw/holdout.csv", index=False)
